+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## Unreleased
 
 - Improved unit test coverage.
+- Added cmdlet `Set-CosmosDBUser` for setting the user Id of an
+  existing user.
 
 ## 1.0.1.40